Create new token generator + update membert gen.
Add AnonymousTokenGenerator and refactor MemberTokenGenerator to use
the same ErrorMessages without code duplication
