+++ src/Helpers/AnonymousTokenGenerator.php
@@ -1,0 +1,31 @@
+<?php
+
+use Firesphere\GraphQLJWT\Resolvers\Resolver;
+
+trait AnonymousTokenGenerator
+{
+  use ErrorMessageGenerator;
+
+  /**
+   * Generate MemberToken response
+   *
+   * @param string $status Status code
+   * @param Member $member
+   * @param string $token
+   * @return array Response in format required by MemberToken
+   */
+  protected static function generateAnonymousResponse(string $status, string $token = null): array
+  {
+    // Success response
+    $valid = $status === Resolver::STATUS_OK;
+    $response = [
+      'valid'   => $valid,
+      'token'   => $token,
+      'status'  => $status,
+      'code'    => $valid ? 200 : 401,
+      'message' => ErrorMessageGenerator::getErrorMessage($status),
+    ];
+
+    return $response;
+  }
+}
