+++ src/Helpers/MemberTokenGenerator.php
@@ -1,7 +1,10 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Firesphere\GraphQLJWT\Helpers;
 
+use ErrorMessageGenerator;
 use Firesphere\GraphQLJWT\Resolvers\Resolver;
 use InvalidArgumentException;
 use SilverStripe\Core\Extensible;
@@ -14,31 +17,6 @@ use SilverStripe\Security\Member;
  */
 trait MemberTokenGenerator
 {
-    /**
-     * Humanise error message based on status code
-     *
-     * @param string $status
-     * @return string
-     * @throws InvalidArgumentException
-     */
-    public static function getErrorMessage(string $status): string
-    {
-        switch ($status) {
-            case Resolver::STATUS_EXPIRED:
-                return _t('JWT.STATUS_EXPIRED', 'Token is expired, please renew your token with a refreshToken query');
-            case Resolver::STATUS_DEAD:
-                return _t('JWT.STATUS_DEAD', 'Token is expired, but is too old to renew. Please log in again.');
-            case Resolver::STATUS_INVALID:
-                return _t('JWT.STATUS_INVALID', 'Invalid token provided');
-            case Resolver::STATUS_BAD_LOGIN:
-                return _t('JWT.STATUS_BAD_LOGIN', 'Sorry your email and password combination is rejected');
-            case Resolver::STATUS_OK:
-                return _t('JWT.STATUS_OK', 'Token is ok');
-            default:
-                throw new InvalidArgumentException("Invalid status");
-        }
-    }
-
     /**
      * Generate MemberToken response
      *
@@ -57,7 +35,7 @@ trait MemberTokenGenerator
             'token'   => $token,
             'status'  => $status,
             'code'    => $valid ? 200 : 401,
-            'message' => static::getErrorMessage($status),
+            'message' => ErrorMessageGenerator::getErrorMessage($status),
         ];
 
         return $response;
